refactor API structure
