+++ .gitignore
@@ -300,4 +300,5 @@ pip-selfcheck.json
 login_logs.txt
 messages.txt
 
+/profile_pics/*
 # End of https://www.toptal.com/developers/gitignore/api/pycharm,django,venv
